ControllerAdvice, ResponseEntity, Projections, Parameter binding

+++ src/main/java/tech/ada/java/movieslibrary/api/movie/MovieDTO.java
@@ -1,6 +1,6 @@
 package tech.ada.java.movieslibrary.api.movie;
 
-public record MovieDTO(String imdbID, String title, String genre, String director, String actors, Long year) {
+public record MovieDTO(String imdbID, String title, String genre, String director, String actors, Integer year) {
 
     public MovieDTO(MovieModel movieModel) {
         this(movieModel.getImdbID(), movieModel.getTitle(), movieModel.getGenre(),
